Reimplement truncate_users without using delete
using instead the TRUNCATE TABLE sql command
